requirements.txt
# Core Streamlit and web framework
streamlit==1.28.1
streamlit-folium==0.15.0

# Map and visualization libraries
folium==0.14.0

# HTTP requests and API calls
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.24.3

# Additional utilities
urllib3==2.0.7